pyproject.toml
[tool.poetry]
name = "django_2fauth"
version = "0.1.0"
description = ""
authors = ["DanyloSolohub <danylo.solohub.te.2019@lpnu.ua>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.9"
django-rest-framework = "^0.1.0"
Pillow = "^8.4.0"
psycopg2-binary = "^2.9.2"
flake8 = "3.8.4"
flake8-quotes = "^3.3.1"
flake8-comprehensions = "^3.7.0"
flake8-absolute-import = "^1.0"
flake8-django = "^1.1.2"
flake8-print = "^4.0.0"
django-cors-headers = "^3.10.0"
django-environ = "^0.8.1"
drf-yasg = "^1.20.0"
djangorestframework-simplejwt = "^5.0.0"
PyQRCode = "^1.2.1"
django-otp = "^1.1.3"
pypng = "^0.0.21"
qrcode = "^7.3.1"
[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
